@@ -1,4 +1,11 @@
-module.exports = {
+export class IOptions {
+  interval: number;
+  waitForCheckInterval: number;
+  customAttribute: string;
+  locators: string;
+}
+
+export const Options: IOptions = {
   interval: 500,
   waitForCheckInterval: 500,
   customAttribute: 'e2e-tag',
