@@ -1,5 +1,5 @@
-module.exports = {
-  FAILED: 'failed',
-  DONE: 'done',
-  INPROGRESS: 'in-progress'
-};
+export enum COMMAND_STATE {
+  FAILED = 'failed',
+  DONE = 'done',
+  INPROGRESS = 'in-progress'
+}
